+++ test/test.geocoder.js
@@ -473,6 +473,47 @@ test('geocoder', function(tt) {
     );
   });
 
+  tt.test('geocoding works correctly around a place with a 0 lat or lng', function(t) {
+    t.plan(1);
+
+    setup({});
+
+    map.setZoom(13);
+    map.setCenter([0, 51.5]);
+
+    geocoder.query('Berlin');
+    geocoder.on(
+      'results',
+      once(function(e) {
+        t.ok(
+          e.features.some((feature) => feature.place_name.indexOf('Berlin') !== -1),
+          'geocoding works correctly around a location with a 0 lat or lng'
+        );
+      })
+    );
+  });
+
+  tt.test('proximity can be set to a value with a 0 lat or lng', function(t) {
+    t.plan(1);
+
+    setup({});
+
+    map.setZoom(13);
+    map.setCenter([32.58, 0]);
+    geocoder.setProximity({ longitude: 32.58, latitude: 0});
+
+    geocoder.query('ent');
+    geocoder.on(
+      'results',
+      once(function(e) {
+        t.ok(
+          e.features[0].place_name.indexOf('Entebbe') !== -1,
+          'proximity applied correctly to location including a zero in geocoding request'
+        );
+      })
+    );
+  });
+
   tt.test('options.render', function(t){
     t.plan(3);
     setup({