+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 ### Bug fixes 🐛
 - disable proximity when reverseGeocoding since it is not supported by the Mapbox Geocoding API [#327](https://github.com/mapbox/mapbox-gl-geocoder/pull/327)
 
+### Bug fixes 🐛
+
+- Fix a bug where the geocoder did not work when used on a map inititalized to a lat or lng of 0
+
 ## v4.5.1
 ### Bug fixes 🐛